Created component library to allow components extending and creation of new components
