+++ lib/inky.js
@@ -61,4 +61,6 @@ Inky.prototype.releaseTheKraken = function($) {
 
 Inky.prototype.componentFactory = require('./componentFactory');
 
+Inky.prototype.componentLibrary = require('./componentLibrary');
+
 Inky.prototype.makeColumn = require('./makeColumn');
